# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Server to get the information about the user \u0026 repo from github using Github Public APIs.",
        "title": "Go-Github Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Show the status of server.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/{github_username}/repos": {
            "get": {
                "description": "Get the User Github Profile and Repo information using public Github APIs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Get the User Github Profile and Repo information.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "string valid",
                        "name": "github_username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.GithubRepoResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.GithubRepoBranchResponse": {
            "type": "object",
            "properties": {
                "branchName": {
                    "type": "string"
                },
                "lastCommitSHA": {
                    "type": "string"
                }
            }
        },
        "model.GithubRepoResponse": {
            "type": "object",
            "properties": {
                "branches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GithubRepoBranchResponse"
                    }
                },
                "ownerLogin": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                }
            }
        }
    }
}